[utils-go][sqldb] Add new sqldb package
sqldb package manages database connections, particularly for postgres.
Migrated fastlinks over to sqldb.
