+++ fastlinks/stores/postgresql/BUILD.bazel
@@ -1,4 +1,4 @@
-load("//bazel/go:default.bzl", "go_library", "go_test")
+load("//bazel/go:default.bzl", "go_library")
 
 go_library(
     name = "postgresql",
@@ -8,16 +8,5 @@ go_library(
     ],
     importpath = "github.com/TerrenceHo/monorepo/fastlinks/stores/postgresql",
     visibility = ["//visibility:public"],
-    deps = [
-        "//utils-go/stackerrors",
-        "@com_github_jmoiron_sqlx//:sqlx",
-        "@com_github_lib_pq//:pq",
-    ],
-)
-
-go_test(
-    name = "postgresql_test",
-    srcs = ["stores_test.go"],
-    embed = [":postgresql"],
-    deps = ["@com_github_stretchr_testify//assert"],
+    deps = ["//utils-go/sqldb"],
 )
